+++ CHANGELOG.md
@@ -6,10 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
-## 2/6/2026
+## 2/26/2026
 
 ### Added
 
-- Documented plans for components in [component brainstorming](doc/01-component-brainstorming/01-component-brainstorming.md) section.
-- FixLater.java added as a placeholder
-- Reccomended extensions and JUnit added to workspace
+- EMNumber class
+    - Methods
+        - mantissa() (getter)
+        - exponent() (getter)
+        - isZero()
+        - add(EMNumber n)
+        - subtract(EMNumber n)
+        - multiply(EMNumber n)
+        - divide(EMNumber n)
+    - Constructors
+        - (double mantissa, int exponent)
+        - (int n)
+        - (double n)
+        - (NaturalNumber n)